feat(API migrations): add 'warranty_requests' table
This commit introduces a new DB table to manage warranty requests. When
a user experiences issues with a purchased product within the warranty
period, they can visit our nearest offline shop. An administrator can
then add a warranty request to the database and transfer the product to
the appropriate service center for diagnostics and repair. The
'warranty_requests' table is used to store such requests.

This table stores the following columns:
- serial_number (serial number of the product that is to be repaired)
- issue_description (complaint from the user)
- user_id (id of the user who came to the shop to repair their product)
- product_id (id of the affected product)
- service_center_id (id of the service center that is responsible for
the repairs)