feat(API migrations): add 'warranty_request_statuses' table
This commit introduces a new DB table to manage statuses of warranty
requests. When a user experiences issues with a purchased product within
the warranty period, they can visit our nearest offline shop. An
administrator can then add a warranty request to the database and
transfer the product to the appropriate service center for diagnostics
and repair. The 'warranty_request_statuses' table is used to store
different statuses for warranty requests. This table has the following
fields:
- id
- name (name of the warranty request status)