feat(API migrations): add 'warranty_request_status_history' table
This commit introduces a new DB table that tracks status changes of
warranty requests. When a user experiences issues with a purchased
product within the warranty period, they can visit our nearest offline
shop. An administrator can then add a warranty request to the database
and transfer the product to the appropriate service center for
diagnostics and repair. The 'warranty_request_status_history' table is
used to store statuses that different warranty requests have at
different points in time.

This table has the following fields:
- warranty_request_id
- status_id (references "warranty_request_statuses"."id")
- change_time (specifies when the status was changed)