feat(API seeds): add a seed file for the 'warranty_request_statuses' table
Seed files are used to prepopulate database tables with sample data for
testing purposes.

This commit adds a file called '17_warranty_request_statuses.ts' that
prepopulates the 'warranty_request_statuses' table. This table stores
statuses of warranty requests. A warranty request is a is a formal
inquiry made by a customer to a seller regarding a product that has
experienced issues covered under the warranty.