feat(API seeds): add a seed file for the 'fundraising_transactions' table
Seed files are used to prepopulate database tables with sample data for
testing purposes.

The file '20_fundraising_transactions.ts' is introduced, populating the
'fundraising_transactions' table. This table records financial
transactions related to fundraising campaigns for military purposes.

The seed file is designed to simulate one campaign as completed, and
two others at around 30-40% progress. For instance, if a campaign aims
to raise 290,000 UAH, approximately 100,000 UAH will be simulated as
raised.

It's important to point out that these transactions are not real, and
are only made for testing purposes.