feat(API): add an endpoint that fetches fundraising campaigns
This commit introduces a new GET /api/fundraising-campaigns endpoint
that returns a list of fundraising campaigns based on the provided
status. The status must be either 'ongoing' or 'finished'.

This endpoint does not support pagination, because this online store is
not expected to run many fundraising campaigns at the same time.

For more details about this endpoint please refer to the API docs.