feat(React): remove admin classes for the top header
Previously, the top header of the React application adapted depending
on the status of the user. This behavior was implemented a long time
ago, and since then the design of the application changed. Now this
application is expected to have a completely different interface for
admins. For now, the top header was modified so that it assumes that the
current user is a basic user. Admin interface will be implemented later.