refactor(React): extract <Subnavigation /> component
This commit refactors the React project by creating a <Subnavigation />
component, which is used for creating navigation links that point to
different pages. This component holds common logic about subnavigation
links.

This component is used as follows:

<SubnavigationItems
    links={[
        { name: 'Orders', url: '/user/orders' },
        {
            name: 'Warranty requests',
            url: '/user/orders/warranty-requests',
        },
    ]}
    selectedLink={'Orders'}
    divBlockClassName={classes['order-subnavigation-items']}
    linkItemClassName={classes['order-subnavigation-items__link']}
/>

The order component that is responsible for the page /user/orders has
been refactored to utilize this new component.