refactor(React): use <Card /> inside <ProductItem />
This commit refactors the <ProductItem /> component to utilize the
<Card /> component. The <Card /> component has common CSS properties
that a typical card would have, such as borders, centered text, shadows,
margins, and more. Because of this refactor, the <ProductItem />
component became more dependable on the Card component, which is a
desired outcome, because the ultimate goal of these refactors is to make
product items and fundraising campaign items dependent on the <Card />
component, which would reduce code duplication.