feat(API): add pending transaction endpoint for campaigns
This commit creates an HTTP endpoint that allows to create a new pending
transaction for the specified fundraising campaign. Pending transactions
are transactions that haven't been paid for yet.

POST /api/fundraising-campaigns/pending-transactions

This endpoint accepts the following parameters in the request body:
- campaignId - must point to the ID of an existing fundraising campaign
- donationAmount - must be a number that is greater than or equal to 100
This endpoint returns "data" and "signature", which are to be used for
redirecting the user to a LiqPay payment page.

For more details about this commit please refer to the API docs.