fix(API): change 'action' when creating a fundraising transaction
When a user tries to contribute to a fundraising campaign, they must
hit the POST /api/fundraising-campaigns/pending-transactions endpoint,
which creates a new pending transaction (one that hasn't been paid for
yet) and returns 'data' and 'signature' parameters, which are to be used
for redirecting users to the LiqPay payment page. The problem is that
this endpoint generates these parameters using the 'paydonate' action,
which indicates that users can donate any amount of money, starting from
0.01 UAH. However, this lower limit cannot be be changed, so the only
way to impose a limit is to use a the 'pay' action, which is a general
action that specifies that this operation is a direct withdrawal from
the card.