refactor(API): transform fundraisingCampaignModel into a class
Subsequent commits will need to use some methods of the fundraising
campaign model inside SQL transactions. To reduce code duplication,
this commit transforms this model into a class, which has a constructor
that accepts a database client, and then all methods inside that
instance will use the same DB client. This will allow methods to share
the same SQL transaction instead of relying on the the pool of DB
connections.

All existing code that relied on this model was refactored to utilize
this newly created class.