feat(API): add endpoint for LiqPay fundraising donation callback
When a user donates money to a fundraising campaign, they will be
redirected to the callback endpoint. This commit introduces this
endpoint, which will accept the 'data' and 'signature' parameters,
validate them and mark the pending transaction is 'paid' if the
parameters indeed came from LiqPay. After this the user will be
redirected to a designated React page.

For more details please refer to the API documentation.