fix(API): prevent transaction creation for finished campaigns
If users wanted to donate money to a fundraising campaign, they would
need to use the POST /api/fundraising-campaigns/pending-transactions
endpoint. This endpoint creates a pending transaction, which is a
transaction that hasn't been paid for yet, and then returns 'data' and
'signature' parameters, which are used to redirect the user to a LiqPay
payment page.

However, this endpoint has a bug: it allowed to create pending
transactions for campaigns that have already raised all the necessary
money. This commit fixes this unwanted behavior by performing a check
in the "createPendingTransaction" controller. Now, if a user attempts
to create a pending transaction for a finished fundraising campaign,
they will get a 409 Conflict error that states "The fundraising campaign
has already reached its goal and is now closed for further donations."

The API documentation has been updated to reflect this new change.

Note: the React application has not been updated to handle this error
yet. This will be addressed in future commits.