feat(React): remove "Home" navigation item
This modification removes the "Home" navigation item from the top menu.
It has been determined that this navigation option is redundant, as
users can easily return to the home page by clicking on the logo located
in the top menu.

Moreover, CSS media queries associated with navigation items have been
adjusted to accommodate this alteration.