fix(API): fix the 'cancelOrder' method of OrderModel
Previously, the 'cancelOrder' method of the 'OrderModel' class was
implemented incorrectly. The previous implementation tried to change
the status of an order by updating the 'status_id' column of the table
called 'orders'. The problem is that the 'orders' table doesn't have
the 'status_id' column. In order to mark the order as canceled, it's
necessary to insert a new row into the 'order_status_history' table,
specifying the cancellation of the order. This commit fixes this bug.