feat(NestJS API): add endpoint for getting OAuth 2.0 consent link
This commit creates an endpoint GET /api/auth/oauth-link/{PROVIDER}
for generating a link to the OAuth 2.0 authorization server. This link
will be used to redirect users to the consent screen. After users
consent, they will be redirected to the callback page. The callback
functionality for NestJS will be introduced in next commits.

Please refer to the Swagger UI for more information about this endpoint.