refactor(NestJS DB seeding): make product category seeding type-safe
Previously, the ProductCategorySeeder class inserted product categories
into the database directly, which is a viable approach, but it's not
great in the long run. It would be better to define a type that stores
all the possible product categories, and then insert these categories
according to this type. This would make the code more maintainable.

Changes:
- Created a separate ProductCategory type that stores all the possible
product categories. This type can be reused across multiple seeders.
Warning: this type is NOT meant to be used in the main codebase. It
should only be used for DB seeding.
- Created a private static property inside the ProductCategorySeeder
class that stores different product categories as an array. Any category
in this array must adhere to the ProductCategory type that was described
above.