fix(NestJS seeding): use 'remove' instead of 'clear'
Previously, the product category seeder file used the "clear" method
to remove all rows inside the "product_categories" table. However, this
method proved to cause errors, particularly when there is a table that
has a foreign key pointing to a product category. Under the hood, the
"clear" method uses the "TRUNCATE" command, which on its own isn't
suitable for clearing a table that is pointed to from another table.
In order to solve this problem, the "delete" function was used instead.