chore(NestJS seeding): add files for product seeding
This adds the ability to pre-populate the "products" table in the DB.
- A new Product TypeORM entity was created, which has a reference to
the product categories table.
- The ProductCategory TypeORM entity was modified to add a relationship
with the newly created Product entity.
- The Product entity was added to 'products.module.ts' in order for
NestJS (and TypeORM in particular) to know about the existence of this
entity.
- A new seeding file ('02-products.ts') was created, which adds predefined
products to the DB.
- The 'seed-runner.ts' file was modified to register this new seeding
file.

From now on, when you run 'npm run seed:run', predefined product
categories as well as different products will be added to the DB for
development and testing purposes.