chore(NestJS seeding): add files for user seeding
This adds the ability to pre-populate the 'users' table using the
seeding mechanism.
To seed the database, run the following command:
$ npm run seed:run

The user seeder class requires some services in order for it to work
correctly. To utilize them, we would need to create a new application
instance and then get the instances of these services. However, this
approach has proved to be infeasible. Instead, a separate isolated
testing module was created in the 'seed-runner.ts' file. This module
includes all the necessary services needed for the seeding process.