feat(Express): migrate to OpenSearch
Previously, the Express.js server utilized Elasticsearch v7.10 to
implement product search via the GraphQL "searchProducts" query. This
version was chosen because it is the latest version supported by the AWS
OpenSearch service.

The problem is that this version is now considered pretty old.
Therefore, there is no alternative other than migrating to OpenSearch,
which is an open-source fork of Elasticsearch initially developed by
Amazon.

OpenSearch is a much better choice for AWS integration. This commit
rewrites the Express codebase to work with OpenSearch v2.15.