feat(NestJS): migrate to OpenSearch
Previously, the Express.js server utilized Elasticsearch v7.10 to
implement product search via the GraphQL "searchProducts" query. This
version was chosen because it is the latest version supported by the AWS
OpenSearch service.

The problem is that this version is now considered pretty old.
Therefore, there is no alternative other than migrating to OpenSearch,
which is an open-source fork of Elasticsearch initially developed by
Amazon. OpenSearch is a much better choice for AWS integration.

This commit rewrites the docker compose file that is associated with the
NestJS app by removing the Elasticsearch service and adding a new
OpenSearch service. In future commits, the NestJS application is
expected to get features that rely heavily on OpenSearch.