feat(Express API): change 'orderId' from SERIAL to UUIDv4
This commit rewrites the 'orders' table by changing type of the 'id'
column from 'SERIAL' to 'UUID' (using migration files). All the relevant
code was modified to account for this change. This only applies to the
Express application, the NestJS and React applications were left as is.

The transition from SERIAL to UUID is necessary because using the type
'SERIAL' complicates the development of the Express application.
Sometimes the PostgreSQL database may be reset, and the order IDs start
from 1, which causes problems with testing the LiqPay payment process
(e.g. trying to pay for the same order twice). Using UUIDs fixes this
problem.

The API documentation was also modified to reflect this change.